Remove Unused Dao (#983)
* Delete Unused Dao

* fix test

* fix lint

* address unused arg
